+++ src/main/java/es/in2/wallet/infrastructure/ebsi/controller/OpenidCredentialOfferController.java
@@ -1,8 +1,8 @@
 package es.in2.wallet.infrastructure.ebsi.controller;
 
+import es.in2.wallet.application.workflow.issuance.CredentialIssuanceCommonWorkflow;
 import es.in2.wallet.infrastructure.core.config.SwaggerConfig;
 import es.in2.wallet.application.workflow.issuance.CredentialIssuanceEbsiWorkflow;
-import es.in2.wallet.domain.model.EbsiCredentialOfferContent;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,28 +19,38 @@ import static es.in2.wallet.domain.util.ApplicationUtils.getCleanBearerToken;
 
 @Slf4j
 @RestController
-@RequestMapping("/api/v1/request-credential")
+@RequestMapping("/api/v1/openid-credential-offer")
 @RequiredArgsConstructor
-public class CredentialIssuanceController {
+public class OpenidCredentialOfferController {
 
     private final CredentialIssuanceEbsiWorkflow ebsiCredentialIssuanceServiceFacade;
+    private final CredentialIssuanceCommonWorkflow commonCredentialIssuanceServiceFacade;
+
 
     /**
      * Processes a request for a verifiable credential when the credential offer is received via a redirect.
      * This endpoint is designed to handle the scenario where a user is redirected to this service with a credential
      * offer URI, as opposed to receiving the offer directly from scanning a QR code.
      */
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
     @Operation(
             tags = (SwaggerConfig.TAG_PUBLIC)
     )
-    public Mono<Void> requestVerifiableCredential(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader,
-                                                                @RequestBody EbsiCredentialOfferContent ebsiCredentialOfferContent) {
+    public Mono<Void> requestOpenidCredentialOffer(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader,
+                                                   @RequestParam String credentialOfferUri) {
         String processId = UUID.randomUUID().toString();
         MDC.put("processId", processId);
         return getCleanBearerToken(authorizationHeader)
-                .flatMap(authorizationToken -> ebsiCredentialIssuanceServiceFacade.identifyAuthMethod(processId, authorizationToken, ebsiCredentialOfferContent.credentialOfferUri()));
+                .flatMap(authorizationToken -> {
+                    if (credentialOfferUri.contains("ebsi")) {
+                        return ebsiCredentialIssuanceServiceFacade.identifyAuthMethod(
+                                processId, authorizationToken, credentialOfferUri);
+                    } else {
+                        return commonCredentialIssuanceServiceFacade.identifyAuthMethod(
+                                processId, authorizationToken, credentialOfferUri);
+                    }
+                });
     }
 
 }